serving/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pandas
Pillow
python-multipart
matplotlib
flask